Try separate state machine from global variables
NOT WORKING!

Suspicion is that the state machine is wasting bytes before getting to a
stable state. Thus, missing commands and failing to enter and process a
valid state and command. The whole structure of how data within the
state machine moves needs to be redesigned to fix some of these
problems.