Add Oli emblem as device greeting
NOTE: STM32 function definitions for UART don't have the `const` specifier
in the transmit function call. I suspect this is a bug since reviewing
their code I can't see a reason why they require it to be a writeable
pointer. As such the PFP have been updated but will be overridden by
CubeMx on next gen.