Fix logging/sai bug
If the device was logging to file and not TTY the SAI interrupt would
not be generated. After a very log search found to be that when logging
to TTY there is a longer delay caused by the synchronous nature of the
TTY logging. This means that some commands don't have time to latch
properly when sent over I2C and thus a manual delay was introduced.
These are the types of bugs that embedded engineers have to deal with...